code 12: Documentation of the Type of Feeding catalog.
