--- src/main/java/com/project/app_dog_back/application/dto/TipoAlimentoDto.java
@@ -1,15 +1,0 @@
-package com.project.app_dog_back.application.dto;
-
-import lombok.Builder;
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-@Builder
-public class TipoAlimentoDto {
-    private Long idTipoAlimento;
-    private String nombre;
-    private String descripcion;
-    private Boolean valido;
-}
