+++ src/main/java/com/project/app_dog_back/application/mapper/ITipoAlimentoMapper.java
@@ -1,6 +1,6 @@
 package com.project.app_dog_back.application.mapper;
 
-import com.project.app_dog_back.application.dto.TipoAlimentoDto;
+import com.project.app_dog_back.application.dto.catalog.TipoAlimentoDto;
 import com.project.app_dog_back.domain.model.catalog.TipoAlimento;
 import org.mapstruct.Mapper;
 import org.mapstruct.Mapping;
